requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
scipy>=1.7.0
Pillow>=8.3.0
matplotlib>=3.4.0
tqdm>=4.62.0
timm>=0.4.12
opencv-python>=4.5.5.64
einops>=0.4.1
# Optional, for NonLocal2d
# mmcv>=1.4.0
